Combine surveys based on diagnosis
